Correção do segundo Bug

+++ desafio-aula-1/registradora-cheia-de-bugs/registradora-cheia-de-bugs/src/ItensPorQuantidade.java
@@ -1,10 +1,11 @@
 public class ItensPorQuantidade {
 
     static int pao = 3600;
-    static int torta = 4;
+    static float torta = 4;
     static int sanduiche = 20;
     static int leite = 20;
     static int cafe = 20;
+
 
 
 }
